Webex form auto-fill and click
The form that asks for the user email is automatically filled and
clicked, so the user gets logged in without taking any action.
An email needs to be provided on the first login attempt, then it will
be saved in localStorage. The email prompt is very rudimental and will
probably be changed with something better in the future.